refactor(api): move archiveCampaigns to prescription campaign bounded context
