--- api/lib/infrastructure/repositories/campaigns-administration/campaign-repository.js
@@ -1,10 +1,0 @@
-import { knex } from '../../../../db/knex-database-connection.js';
-
-const archiveCampaigns = function (campaignIds, userId) {
-  return knex('campaigns').whereNull('archivedAt').whereInArray('id', campaignIds).update({
-    archivedBy: userId,
-    archivedAt: new Date(),
-  });
-};
-
-export { archiveCampaigns };
